// SDKs/onsched.com/consumer/v1/typescript/src/sdk/models/shared/addressupdatemodel.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class AddressUpdateModel extends SpeakeasyBase {
  @Metadata({ data: "json, name=addressLine1" })
  addressLine1?: string;

  @Metadata({ data: "json, name=addressLine2" })
  addressLine2?: string;

  @Metadata({ data: "json, name=city" })
  city?: string;

  @Metadata({ data: "json, name=country" })
  country?: string;

  @Metadata({ data: "json, name=postalCode" })
  postalCode?: string;

  @Metadata({ data: "json, name=state" })
  state?: string;
}